Move PageMenu component into PageSet pattern, and update to use it.
