+++ package.json
@@ -32,7 +32,7 @@
     "turbo": "^1.12.3",
     "type-fest": "^4.10.2",
     "typescript": "^5.3.3",
-    "vitest": "^0.34.6",
+    "vitest": "^1.6.0",
     "vitest-github-actions-reporter": "^0.11.1",
     "vitest-mock-extended": "^1.3.1"
   },
