+++ packages/design/package.json
@@ -13,11 +13,10 @@
     "clean:styles": "rm -rf static",
     "dev": "run-p dev:*",
     "dev:lib": "vite",
-    "dev:storybook": "storybook dev",
+    "dev:storybook": "storybook dev -p 9009",
     "dev:styles": "gulp watch",
     "lint": "eslint --ext ts,tsx --report-unused-disable-directives --max-warnings 0 src",
-    "test:storybook": "test-storybook",
-    "test:url": "test-storybook --url http://127.0.0.1:9009 --config-dir storybook",
+    "test:url": "test-storybook --url http://127.0.0.1:9009 --config-dir .storybook",
     "test:ci": "concurrently -k -s first -n \"SB,TEST\" -c \"magenta,blue\" \"pnpm build:storybook --quiet && npx http-server storybook/public --port 9009 --silent\" \"wait-on tcp:127.0.0.1:9009 && pnpm test:url --maxWorkers=2\""
   },
   "files": [
@@ -37,14 +36,14 @@
     "@storybook/test": "^7.6.10",
     "@storybook/test-runner": "^0.16.0",
     "@storybook/types": "^7.6.10",
-    "@testing-library/react": "^15.0.2",
+    "@testing-library/react": "^15.0.7",
     "@types/deep-equal": "^1.0.4",
     "@types/prop-types": "^15.7.12",
     "@types/react": "^18.2.79",
     "@typescript-eslint/eslint-plugin": "^7.7.0",
     "@typescript-eslint/parser": "^7.7.0",
     "@uswds/compile": "1.1.0",
-    "@vitejs/plugin-react": "^4.2.1",
+    "@vitejs/plugin-react": "^4.3.0",
     "concurrently": "^8.2.2",
     "eslint": "^8.56.0",
     "eslint-plugin-react": "^7.34.1",
@@ -54,8 +53,8 @@
     "install": "^0.13.0",
     "prop-types": "^15.8.1",
     "react-dom": "^18.2.0",
-    "vite": "^5.2.9",
-    "vite-plugin-dts": "^3.8.3",
+    "vite": "^5.2.11",
+    "vite-plugin-dts": "^3.9.1",
     "wait-on": "^7.2.0"
   },
   "dependencies": {
