+++ packages/design/src/Form/components/PageSet/PageMenu/PageMenu.test.tsx
@@ -1,7 +1,7 @@
 /**
  * @vitest-environment jsdom
  */
-import { describeStories } from '../../../test-helper';
+import { describeStories } from '../../../../test-helper';
 import meta, * as stories from './PageMenu.stories';
 
 describeStories(meta, stories);
