+++ packages/design/src/Form/components/PageSet/PageSet.tsx
@@ -6,6 +6,7 @@ import { type PageSetProps } from '@atj/forms';
 
 import { type PatternComponent } from '../..';
 import { useFormManagerStore } from '../../../FormManager/store';
+import { PageMenu } from './PageMenu';
 
 const PageSet: PatternComponent<PageSetProps> = props => {
   const location = useLocation();
@@ -13,24 +14,17 @@ const PageSet: PatternComponent<PageSetProps> = props => {
   return (
     <div className="grid-row grid-gap">
       <nav className="tablet:grid-col-3 bg-primary-lightest">
-        <ul className="usa-sidenav">
-          {props.pages.map((page, index) => {
+        <PageMenu
+          pages={props.pages.map((page, index) => {
             const params = new URLSearchParams(routeParams?.toString());
             params.set('page', index.toString());
-            return (
-              <li key={index} className={'usa-sidenav__item'}>
-                <a
-                  className={classNames({
-                    'usa-current': page.active,
-                  })}
-                  href={`#${location.pathname}?page=${index}`}
-                >
-                  {page.title}
-                </a>
-              </li>
-            );
+            return {
+              title: page.title,
+              selected: page.active,
+              url: `#${location.pathname}?page=${index}`,
+            };
           })}
-        </ul>
+        />
       </nav>
       <div className="tablet:grid-col-9">{props.children}</div>
     </div>
