+++ packages/design/src/FormManager/FormEdit/components/PageSetEdit.stories.tsx
@@ -1,5 +1,3 @@
-import React from 'react';
-import { MemoryRouter } from 'react-router-dom';
 import type { Meta, StoryObj } from '@storybook/react';
 import { within } from '@testing-library/react';
