+++ packages/docassemble/package.json
@@ -15,6 +15,6 @@
   },
   "devDependencies": {
     "@vitest/coverage-v8": "^0.34.6",
-    "vitest": "^0.34.6"
+    "vitest": "^1.6.0"
   }
 }
